Extract data interoplation logic to separate files
